refactor: webhook policy context creation
